+++ src/main/settings/general-settings.ts
@@ -17,7 +17,7 @@ import {
 
 import { GENERAL_SETTINGS_SCHEMA, GeneralSettings } from '../../common/settings-schemata';
 
-import { Settings } from './settings';
+import { getConfigDirectory, Settings } from '.';
 
 import { version } from './../../../package.json';
 
@@ -33,6 +33,7 @@ export function getGeneralSettings(): Settings<GeneralSettings> | null {
   try {
     return new Settings<GeneralSettings>({
       file: 'config.json',
+      directory: getConfigDirectory(),
       defaults: () => GENERAL_SETTINGS_SCHEMA.parse({}),
       load: (content) => loadGeneralSettings(content),
     });
